No idea whats up here...
git-svn-id: https://wokjab.svn.sourceforge.net/svnroot/wokjab/wokjab/trunk/wokjab@240 a4058eb7-c820-0410-9b0b-94ec00e1a786

+++ plugins/experimental/Dice/dice-session.cpp
@@ -174,15 +174,15 @@ DiceSession::Message(WokXMLTag *tag)
 int 
 DiceSession::RandNR(WokXMLTag *tag)
 {	
-	long unsigned int num = strtoul(("0x" + tag->GetAttr("num")).c_str(), NULL, 16);
-	long unsigned int bits = 0xFFFFFFFFFFFFFFFF;
+	long long unsigned int num = strtoul(("0x" + tag->GetAttr("num")).c_str(), NULL, 16);
+	long long unsigned int bits = 0xFFFFFFFFFFFFFFFF;
 		
 	std::stringstream result_str;
 	
 	std::list <WokXMLTag *>::iterator dieiter;
 	for( dieiter = dice_data[tag->GetAttr("owner")]->GetTagList("die").begin(); dieiter != dice_data[tag->GetAttr("owner")]->GetTagList("die").end() ; dieiter++)
 	{
-		long unsigned int faces = atoi((*dieiter)->GetFirstTag("type", "common notation").GetBody().substr(1).c_str());
+		long long unsigned int faces = atoi((*dieiter)->GetFirstTag("type", "common notation").GetBody().substr(1).c_str());
 		
 		if ( faces > bits )
 		{
@@ -190,7 +190,7 @@ DiceSession::RandNR(WokXMLTag *tag)
 			break;
 		}
 		
-		long unsigned int result = num%faces + 1;
+		long long unsigned int result = num%faces + 1;
 		num /= faces;
 		bits /= faces;
